+++ tests/app/urls.py
@@ -1,9 +1,6 @@
-from django.conf.urls import include, url
-from wagtail.core import urls as wagtail_urls
-
-from wagtailformblocks import urls as formblocks_urls
+from django.urls import include, path
 
 urlpatterns = [
-    url(r'^forms/', include(formblocks_urls)),
-    url(r'', include(wagtail_urls)),
+    path('forms/', include('wagtailformblocks.urls')),
+    path('', include('wagtail.core.urls')),
 ]