+++ wagtailformblocks/utils/conf.py
@@ -1,11 +1,5 @@
-from django import VERSION as DJANGO_VERSION
 from django.conf import settings
-
-if DJANGO_VERSION < (3, 0):
-    from django.utils.translation import ugettext_lazy as _
-else:
-    from django.utils.translation import gettext_lazy as _
-
+from django.utils.translation import gettext_lazy as _
 
 DEFAULTS = {
     'SUCCESS_MSG': _('Thank you, the form has been submitted.'),
@@ -15,8 +9,7 @@ DEFAULTS = {
 
 
 def get_formblocks_setting(name):
-    return getattr(
-        settings, 'WAGTAIL_FORMBLOCKS_{}'.format(name), DEFAULTS[name])
+    return getattr(settings, 'WAGTAIL_FORMBLOCKS_{}'.format(name), DEFAULTS[name])
 
 
 def recaptcha_enabled():